% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallory_predictor.R
\name{mallory_predictor}
\alias{mallory_predictor}
\title{Title}
\usage{
mallory_predictor(person, time_of_day)
}
\arguments{
\item{person}{(lewis or victoria)}

\item{time_of_day}{(morning, afternoon, or evening)}
}
\description{
Title
}
\examples{
mallory_predictor("lewis", "afternoon")
}
