% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_price.R
\name{nail_price}
\alias{nail_price}
\title{calculate the price of getting your nails done}
\usage{
nail_price(nailpolish_price, nailart_price)
}
\arguments{
\item{nailpolish_price}{price of your nail polish (either gel or regular, in $)}

\item{nailart_price}{price of nail art}
}
\description{
calculate the price of getting your nails done
}
\examples{
nail_price(nailpolish_price = 40, nailart_price = 20)
}
